# api-client/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Decorating Platform API",
    "description" : "Quoting and customer modules.",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "/",
    "description" : "Same origin"
  } ],
...
  "paths" : {
...
    "/api/quote-requests/{id}/submit" : {
      "post" : {
        "tags" : [ "quote-request-controller" ],
        "summary" : "Hand the request over: every frame is in and the phone is verified",
        "operationId" : "submit",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "The draft this session owns",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Accepted for analysis, with when to expect an answer",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SubmissionResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "No session cookie",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SubmissionResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "The session owns a different draft",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SubmissionResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "Frames are missing, the phone is not verified, or it was already sent",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SubmissionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/quote-requests/{id}/rooms/confirm" : {
      "post" : {
        "tags" : [ "quote-request-controller" ],
        "summary" : "Accept the areas to photograph",
        "operationId" : "confirmRooms",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "The draft this session owns",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ConfirmRoomsRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "The agreed areas, labelled and ordered, with the frames each needs",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConfirmedRoomsResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "An empty or impossible list",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConfirmedRoomsResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "No session cookie",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConfirmedRoomsResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "The session owns a different draft",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConfirmedRoomsResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "Already confirmed",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConfirmedRoomsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/quote-requests/{id}/estimate" : {
      "post" : {
        "tags" : [ "quote-request-controller" ],
        "summary" : "The instant range for the answers given so far",
        "operationId" : "estimate",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "The draft this session owns",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The range, the areas it assumed, and how wide the band is",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StageOneEstimateResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "No session cookie",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StageOneEstimateResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "The session owns a different draft",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StageOneEstimateResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "Not all of §2.1's questions are answered yet",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StageOneEstimateResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/quote-requests/{id}/estimate-sms" : {
      "post" : {
        "tags" : [ "quote-request-controller" ],
        "summary" : "Send the range to a phone, and keep the number",
        "operationId" : "sendEstimateSms",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "The draft this session owns",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SendEstimateSmsRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "202" : {
            "description" : "Accepted. Queued — no provider is configured yet, so nothing claims to have been delivered."
          },
          "400" : {
            "description" : "Not a Turkish mobile number"
          },
          "401" : {
            "description" : "No session cookie"
          },
          "403" : {
            "description" : "The session owns a different draft"
          },
          "409" : {
            "description" : "No range has been computed yet"
          }
        }
      }
    },
    "/api/quote-requests/{id}/consents" : {
      "post" : {
        "tags" : [ "quote-request-controller" ],
        "summary" : "Record the customer's decision on the data notice",
        "operationId" : "consent",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "The draft this session owns",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RecordConsentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "The decision, as it was recorded",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConsentResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Not a decision this API accepts",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConsentResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "No session cookie",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConsentResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "The session owns a different draft",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConsentResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "The notice has changed, or the room list is not agreed yet",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConsentResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/quote-requests/{id}" : {
      "get" : {
        "tags" : [ "quote-request-controller" ],
        "summary" : "The draft as stored, for a screen that has to survive a reload",
        "operationId" : "show",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "The draft this session owns",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The draft",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/QuoteRequestResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "No session cookie",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/QuoteRequestResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "The session owns a different draft",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/QuoteRequestResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "No such draft",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/QuoteRequestResponse"
                }
              }
            }
          }
        }
      },
      "patch" : {
        "tags" : [ "quote-request-controller" ],
        "summary" : "Answer more of the eight questions",
        "operationId" : "answer",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "The draft this session owns",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PatchQuoteRequestRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "The draft as stored, after the merge",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/QuoteRequestResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "No session cookie",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/QuoteRequestResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "The session owns a different draft",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/QuoteRequestResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "No longer a draft: the answers are fixed once the room list is confirmed",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/QuoteRequestResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/quote-requests/{id}/rooms" : {
      "get" : {
        "tags" : [ "quote-request-controller" ],
        "summary" : "The areas to photograph and how far the capture has got",
        "operationId" : "capture",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "The draft this session owns",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The agreed areas, their frames, and which of them have arrived",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CaptureStateResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "No session cookie",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CaptureStateResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "The session owns a different draft",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CaptureStateResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/quote-requests/resume/{token}" : {
      "get" : {
        "tags" : [ "quote-request-controller" ],
        "summary" : "Exchange a handoff token for a session on this device",
        "operationId" : "resume",
        "parameters" : [ {
          "name" : "token",
          "in" : "path",
          "description" : "From an SMS link or a QR code",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The draft, with a session cookie for this device",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/QuoteRequestResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Unknown or expired token",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/QuoteRequestResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "SubmissionResponse" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "string",
            "description" : "ANALYZING"
          },
          "respondBy" : {
            "type" : "string",
            "format" : "date-time",
            "description" : "When the customer is told to expect an answer, computed against §8's hours"
          }
        },
        "required" : [ "respondBy", "status" ]
      },
      "ConfirmRoomsRequest" : {
        "type" : "object",
        "properties" : {
          "areas" : {
            "type" : "array",
            "example" : [ "LIVING_ROOM", "BEDROOM", "BEDROOM", "KITCHEN", "BATHROOM" ],
            "items" : {
              "type" : "string",
              "enum" : [ "LIVING_ROOM", "MASTER_BEDROOM", "BEDROOM", "STUDY", "KITCHEN", "BATHROOM", "HALLWAY", "BALCONY" ]
            },
            "minItems" : 1
          }
        },
        "required" : [ "areas" ]
      },
      "ConfirmedRoomsResponse" : {
        "type" : "object",
        "properties" : {
          "rooms" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Room"
            }
          },
          "photoCount" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "required" : [ "photoCount", "rooms" ]
      },
      "Room" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "type" : {
            "type" : "string",
            "enum" : [ "LIVING_ROOM", "MASTER_BEDROOM", "BEDROOM", "STUDY", "KITCHEN", "BATHROOM", "HALLWAY", "BALCONY" ]
          },
          "label" : {
            "type" : "string"
          },
          "requiredPhotos" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "enum" : [ "WALL_1", "WALL_2", "WALL_3", "WALL_4", "CEILING", "DETAIL" ]
            }
          },
          "captureComplete" : {
            "type" : "boolean"
          }
        },
        "required" : [ "captureComplete", "id", "label", "requiredPhotos", "type" ]
      },
...
      "StageOneEstimateResponse" : {
        "type" : "object",
        "properties" : {
          "low" : {
            "type" : "number"
          },
          "high" : {
            "type" : "number"
          },
          "bandRatio" : {
            "type" : "number"
          },
          "netArea" : {
            "type" : "number"
          },
          "areaWasGross" : {
            "type" : "boolean"
          },
          "rooms" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CalculatedRoomResponse"
            }
          },
          "photoCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "requiredPhotosByType" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "integer",
              "format" : "int32"
            },
            "description" : "How many frames each kind of area needs, in the version that priced this range. §2.2 lets the customer add an area the layout did not imply."
          }
        },
        "required" : [ "areaWasGross", "bandRatio", "high", "low", "netArea", "photoCount", "requiredPhotosByType", "rooms" ]
      },
      "SendEstimateSmsRequest" : {
        "type" : "object",
        "properties" : {
          "phone" : {
            "type" : "string",
            "example" : "0555 123 45 67",
            "minLength" : 1
          }
        },
        "required" : [ "phone" ]
      },
      "RecordConsentRequest" : {
        "type" : "object",
        "properties" : {
          "type" : {
            "type" : "string",
            "enum" : [ "PROCESSING", "RETENTION_FOR_IMPROVEMENT" ],
            "example" : "PROCESSING"
          },
          "granted" : {
            "type" : "boolean"
          },
          "textVersion" : {
            "type" : "string",
            "description" : "The version served by GET /api/consent-notices/{type}",
            "example" : "v1",
            "minLength" : 1
          }
        },
        "required" : [ "granted", "textVersion", "type" ]
      },
      "ConsentResponse" : {
        "type" : "object",
        "properties" : {
          "type" : {
            "type" : "string",
            "enum" : [ "PROCESSING", "RETENTION_FOR_IMPROVEMENT" ]
          },
          "granted" : {
            "type" : "boolean"
          },
          "textVersion" : {
            "type" : "string",
            "example" : "v1"
          },
          "recordedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "required" : [ "granted", "recordedAt", "textVersion", "type" ]
      },
...
      "CaptureStateResponse" : {
        "type" : "object",
        "properties" : {
          "areas" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CaptureAreaResponse"
            }
          },
          "required" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Frames §2.4 asks for in total"
          },
          "taken" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Frames that have arrived"
          },
          "complete" : {
            "type" : "boolean"
          }
        },
        "required" : [ "areas", "complete", "required", "taken" ]
      },
...
    }
  }
}