# api/src/main/java/com/burakpadr/decorating/quoting/adapter/in/web/QuoteRequestController.java
package com.burakpadr.decorating.quoting.adapter.in.web;

import com.burakpadr.decorating.config.session.AnonymousSessionCookie;
import com.burakpadr.decorating.config.session.OwnedQuoteRequest;
import com.burakpadr.decorating.quoting.domain.model.QuoteRequest;
import com.burakpadr.decorating.quoting.domain.port.in.ConfirmRoomList;
import com.burakpadr.decorating.quoting.domain.port.in.ReadCaptureState;
import com.burakpadr.decorating.quoting.domain.port.in.SubmitQuoteRequest;
import com.burakpadr.decorating.quoting.domain.port.in.RecordConsent;
import com.burakpadr.decorating.quoting.domain.port.in.EstimateStageOne;
import com.burakpadr.decorating.quoting.domain.port.in.ManageQuoteRequests;
import com.burakpadr.decorating.quoting.domain.model.QuoteRequestNotFound;
import com.burakpadr.decorating.quoting.domain.port.in.SendEstimateBySms;
import com.burakpadr.decorating.quoting.domain.port.out.ResumeTokens;
import com.burakpadr.decorating.shared.PhoneNumber;
import java.util.UUID;
import org.springframework.web.bind.annotation.PathVariable;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import jakarta.validation.Valid;
import org.springframework.http.HttpHeaders;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PatchMapping;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

/**
 * Starting and filling in a stage 1 draft (§7, BOYA-25).
 *
 * <p>Anonymous: §4.1 forbids a customer row before OTP verification, so the only thing tying a browser
 * to its draft is the session cookie this controller issues (BOYA-24). The POST is the one route here
 * without one — it is what hands it out.
 *
 * <p>Workflow §8 is why PATCH exists rather than a single submit at the end: people fill in two screens
 * on a laptop and finish on a phone, and an answer that lived only in the browser is an answer they get
 * asked for twice.
 *
 * <p>The {@code {id}} path parameters below are declared for the document rather than for the handler.
 * {@link OwnedQuoteRequest} arrives from an argument resolver, so springdoc sees a parameter it cannot
 * classify and — left alone — publishes the resolver's own type as a required query parameter while
 * never mentioning the path variable at all. Both halves are wrong the same way: the generated client
 * ends up unable to call the endpoint it was generated for. {@code OpenApiContractTest} is what
 * notices.
 */
@RestController
@RequestMapping("/api/quote-requests")
class QuoteRequestController {

	private final ManageQuoteRequests requests;
	private final EstimateStageOne estimates;
	private final SendEstimateBySms sms;
	private final ConfirmRoomList roomList;
	private final RecordConsent consents;
	private final ReadCaptureState capture;
	private final SubmitQuoteRequest submission;
	private final ResumeTokens resumeTokens;
	private final AnonymousSessionCookie session;

	QuoteRequestController(ManageQuoteRequests requests, EstimateStageOne estimates,
			SendEstimateBySms sms, ConfirmRoomList roomList, RecordConsent consents,
			ReadCaptureState capture, SubmitQuoteRequest submission, ResumeTokens resumeTokens,
			AnonymousSessionCookie session) {
		this.requests = requests;
		this.estimates = estimates;
		this.sms = sms;
		this.roomList = roomList;
		this.consents = consents;
		this.capture = capture;
		this.submission = submission;
		this.resumeTokens = resumeTokens;
		this.session = session;
	}

	@PostMapping
	@Operation(summary = "Start a draft and take ownership of it")
	@ApiResponses({
			@ApiResponse(responseCode = "201",
					description = "Created. Carries the session cookie every later call needs.")})
	ResponseEntity<QuoteRequestResponse> create() {
		QuoteRequest draft = requests.createDraft();
		// The cookie is the response, as much as the body is: a caller that stored the id and dropped the
		// Set-Cookie header would hold an id it can never use again.
		return ResponseEntity.status(HttpStatus.CREATED)
				.header(HttpHeaders.SET_COOKIE, session.asCookie(draft.id()).toString())
				.body(QuoteRequestResponse.of(draft));
	}

	@GetMapping("/{id}")
	@Operation(summary = "The draft as stored, for a screen that has to survive a reload")
	@Parameter(name = "id", in = ParameterIn.PATH, required = true,
			description = "The draft this session owns",
			schema = @Schema(type = "string", format = "uuid"))
	@ApiResponses({
			@ApiResponse(responseCode = "200", description = "The draft"),
			@ApiResponse(responseCode = "401", description = "No session cookie", content = {}),
			@ApiResponse(responseCode = "403", description = "The session owns a different draft",
					content = {}),
			@ApiResponse(responseCode = "404", description = "No such draft", content = {})})
	QuoteRequestResponse show(@Parameter(hidden = true) OwnedQuoteRequest owned) {
		// §7 does not list this route, and the result screen cannot work without it: it shows the answers
		// back to the customer (workflow §1.5) and has to survive a reload, a back button and the SMS
		// resume link. The alternative is carrying the answers in client state, which fails at exactly the
		// moment §8 cares about — the customer who comes back on another device. Same reasoning as
		// docs/decisions/0015.
		return QuoteRequestResponse.of(requests.find(owned.id()));
	}

	@PatchMapping("/{id}")
	@Operation(summary = "Answer more of the eight questions")
	@Parameter(name = "id", in = ParameterIn.PATH, required = true,
			description = "The draft this session owns",
			schema = @Schema(type = "string", format = "uuid"))
	@ApiResponses({
			@ApiResponse(responseCode = "200", description = "The draft as stored, after the merge"),
			@ApiResponse(responseCode = "401", description = "No session cookie", content = {}),
			@ApiResponse(responseCode = "403", description = "The session owns a different draft",
					content = {}),
			@ApiResponse(responseCode = "409",
					description = "No longer a draft: the answers are fixed once the room list is "
							+ "confirmed", content = {})})
	QuoteRequestResponse answer(
			// Not a @PathVariable. The type is only obtainable from the resolver that has already matched
			// the cookie against this path's id (BOYA-24), so the check cannot be forgotten here — and
			// hidden from the document, because it is resolved rather than sent.
			@Parameter(hidden = true) OwnedQuoteRequest owned,
			@Valid @RequestBody PatchQuoteRequestRequest request) {
		return QuoteRequestResponse.of(requests.answer(owned.id(), request.toAnswers()));
	}

	@PostMapping("/{id}/estimate-sms")
	@Operation(summary = "Send the range to a phone, and keep the number")
	@Parameter(name = "id", in = ParameterIn.PATH, required = true,
			description = "The draft this session owns",
			schema = @Schema(type = "string", format = "uuid"))
	@ApiResponses({
			@ApiResponse(responseCode = "202",
					description = "Accepted. Queued — no provider is configured yet, so nothing claims to "
							+ "have been delivered.", content = {}),
			@ApiResponse(responseCode = "400", description = "Not a Turkish mobile number", content = {}),
			@ApiResponse(responseCode = "401", description = "No session cookie", content = {}),
			@ApiResponse(responseCode = "403", description = "The session owns a different draft",
					content = {}),
			@ApiResponse(responseCode = "409", description = "No range has been computed yet",
					content = {})})
	ResponseEntity<Void> sendEstimateSms(@Parameter(hidden = true) OwnedQuoteRequest owned,
			@Valid @RequestBody SendEstimateSmsRequest request) {
		// 202 rather than 200: the message is queued and, until BOYA-6 picks a provider, that is where it
		// stays. Answering 200 would tell the browser something was delivered.
		sms.send(owned.id(), PhoneNumber.of(request.phone()));
		return ResponseEntity.accepted().build();
	}

	@PostMapping("/{id}/rooms/confirm")
	@Operation(summary = "Accept the areas to photograph")
	@Parameter(name = "id", in = ParameterIn.PATH, required = true,
			description = "The draft this session owns",
			schema = @Schema(type = "string", format = "uuid"))
	@ApiResponses({
			@ApiResponse(responseCode = "200",
					description = "The agreed areas, labelled and ordered, with the frames each needs"),
			@ApiResponse(responseCode = "400", description = "An empty or impossible list", content = {}),
			@ApiResponse(responseCode = "401", description = "No session cookie", content = {}),
			@ApiResponse(responseCode = "403", description = "The session owns a different draft",
					content = {}),
			@ApiResponse(responseCode = "409", description = "Already confirmed", content = {})})
	ConfirmedRoomsResponse confirmRooms(@Parameter(hidden = true) OwnedQuoteRequest owned,
			@Valid @RequestBody ConfirmRoomsRequest request) {
		return ConfirmedRoomsResponse.of(roomList.confirm(owned.id(), request.areas()));
	}

	@PostMapping("/{id}/consents")
	@Operation(summary = "Record the customer's decision on the data notice")
	@Parameter(name = "id", in = ParameterIn.PATH, required = true,
			description = "The draft this session owns",
			schema = @Schema(type = "string", format = "uuid"))
	@ApiResponses({
			@ApiResponse(responseCode = "201", description = "The decision, as it was recorded"),
			@ApiResponse(responseCode = "400", description = "Not a decision this API accepts",
					content = {}),
			@ApiResponse(responseCode = "401", description = "No session cookie", content = {}),
			@ApiResponse(responseCode = "403", description = "The session owns a different draft",
					content = {}),
			@ApiResponse(responseCode = "409",
					description = "The notice has changed, or the room list is not agreed yet",
					content = {})})
	ResponseEntity<ConsentResponse> consent(@Parameter(hidden = true) OwnedQuoteRequest owned,
			@Valid @RequestBody RecordConsentRequest request) {
		// 201 rather than 200: a decision is appended, never updated (§12 keeps every one), so each call
		// that succeeds has created something.
		return ResponseEntity.status(HttpStatus.CREATED).body(ConsentResponse.of(consents.record(
				owned.id(), request.type(), request.granted(), request.textVersion())));
	}

	@GetMapping("/{id}/rooms")
	@Operation(summary = "The areas to photograph and how far the capture has got")
	@Parameter(name = "id", in = ParameterIn.PATH, required = true,
			description = "The draft this session owns",
			schema = @Schema(type = "string", format = "uuid"))
	@ApiResponses({
			@ApiResponse(responseCode = "200",
					description = "The agreed areas, their frames, and which of them have arrived"),
			@ApiResponse(responseCode = "401", description = "No session cookie", content = {}),
			@ApiResponse(responseCode = "403", description = "The session owns a different draft",
					content = {})})
	CaptureStateResponse capture(@Parameter(hidden = true) OwnedQuoteRequest owned) {
		// The capture screen opens on this and on nothing else: it can be reached on a phone that has
		// never seen the room list, which is what the QR handoff is for (§10, BOYA-35).
		return CaptureStateResponse.of(capture.of(owned.id()));
	}

	@PostMapping("/{id}/submit")
	@Operation(summary = "Hand the request over: every frame is in and the phone is verified")
	@Parameter(name = "id", in = ParameterIn.PATH, required = true,
			description = "The draft this session owns",
			schema = @Schema(type = "string", format = "uuid"))
	@ApiResponses({
			@ApiResponse(responseCode = "200",
					description = "Accepted for analysis, with when to expect an answer"),
			@ApiResponse(responseCode = "401", description = "No session cookie", content = {}),
			@ApiResponse(responseCode = "403", description = "The session owns a different draft",
					content = {}),
			@ApiResponse(responseCode = "409",
					description = "Frames are missing, the phone is not verified, or it was already sent",
					content = {})})
	SubmissionResponse submit(@Parameter(hidden = true) OwnedQuoteRequest owned) {
		return SubmissionResponse.of(submission.submit(owned.id()));
	}

	@GetMapping("/resume/{token}")
	@Operation(summary = "Exchange a handoff token for a session on this device")
	@Parameter(name = "token", in = ParameterIn.PATH, required = true,
			description = "From an SMS link or a QR code", schema = @Schema(type = "string"))
	@ApiResponses({
			@ApiResponse(responseCode = "200",
					description = "The draft, with a session cookie for this device"),
			@ApiResponse(responseCode = "404", description = "Unknown or expired token", content = {})})
	ResponseEntity<QuoteRequestResponse> resume(@PathVariable String token) {
		// The one route here that grants a session rather than requiring one — which is its whole purpose:
		// the cookie is on the laptop and the customer is holding a phone. It takes no OwnedQuoteRequest
		// for the same reason, and the token is the credential (§7).
		UUID id = resumeTokens.resolve(token)
				.orElseThrow(() -> new QuoteRequestNotFound(token));
		return ResponseEntity.ok()
				.header(HttpHeaders.SET_COOKIE, session.asCookie(id).toString())
				.body(QuoteRequestResponse.of(requests.find(id)));
	}

	@PostMapping("/{id}/estimate")
	@Operation(summary = "The instant range for the answers given so far")
	@Parameter(name = "id", in = ParameterIn.PATH, required = true,
			description = "The draft this session owns",
			schema = @Schema(type = "string", format = "uuid"))
	@ApiResponses({
			@ApiResponse(responseCode = "200",
					description = "The range, the areas it assumed, and how wide the band is"),
			@ApiResponse(responseCode = "401", description = "No session cookie", content = {}),
			@ApiResponse(responseCode = "403", description = "The session owns a different draft",
					content = {}),
			@ApiResponse(responseCode = "409",
					description = "Not all of §2.1's questions are answered yet", content = {})})
	StageOneEstimateResponse estimate(@Parameter(hidden = true) OwnedQuoteRequest owned) {
		return StageOneEstimateResponse.of(estimates.estimate(owned.id()));
	}
}